feat(historic): add confirm dialog before delete a match

+++ src/app/components/historic/components/match-list/match-list.component.ts
@@ -1,7 +1,7 @@
-import {Component, Inject, Input, OnInit, ViewEncapsulation} from '@angular/core';
+import {Component, inject, Inject, Input, OnInit, ViewEncapsulation} from '@angular/core';
 import {Chip} from "primeng/chip";
 import {NgForOf} from "@angular/common";
-import {PrimeTemplate} from "primeng/api";
+import {ConfirmationService, PrimeTemplate} from "primeng/api";
 import {Skeleton} from "primeng/skeleton";
 import {TableModule} from "primeng/table";
 import {Match} from '../../../../types/match.type';
@@ -9,19 +9,22 @@ import {PaginatedRequest} from '../../../../types/pagination.type';
 import {ApiMatchInterface} from '../../../../services/api-match/api-match.interface';
 import {SessionStat} from '../../../../types/session-stat.type';
 import {Button} from 'primeng/button';
+import {ConfirmDialog} from 'primeng/confirmdialog';
 
 @Component({
-  selector: 'app-match-list',
+    selector: 'app-match-list',
     imports: [
         Chip,
         NgForOf,
         PrimeTemplate,
         Skeleton,
         TableModule,
-        Button
+        Button,
+        ConfirmDialog
     ],
-  templateUrl: './match-list.component.html',
-  styleUrl: './match-list.component.css',
+    providers: [ConfirmationService],
+    templateUrl: './match-list.component.html',
+    styleUrl: './match-list.component.css',
     encapsulation: ViewEncapsulation.None
 })
 export class MatchListComponent implements OnInit {
@@ -33,6 +36,7 @@ export class MatchListComponent implements OnInit {
 
     constructor(
         @Inject('ApiMatchInterface') private readonly apiMatchService: ApiMatchInterface,
+        private readonly confirmationService: ConfirmationService
     ) {
     };
 
@@ -46,11 +50,18 @@ export class MatchListComponent implements OnInit {
         )
     }
 
-    onDeleteMatch (matchId: number) {
-        this.apiMatchService.deleteMatch(matchId).subscribe(
-            () => {
-                this.matches = this.matches.filter(match => match.id !== matchId);
+    onDeleteMatch(matchId: number) {
+        this.confirmationService.confirm({
+            message: 'Cette action est irréversible.',
+            header: 'Supprimer le match ? ',
+            icon: 'pi pi-exclamation-triangle',
+            accept: () => {
+                this.apiMatchService.deleteMatch(matchId).subscribe(
+                    () => {
+                        this.matches = this.matches.filter(match => match.id !== matchId);
+                    }
+                )
             }
-        )
+        });
     }
 }
